// src/lib/types/database.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.5"
  }
  public: {
    Tables: {
// ...
      prompt_templates: {
        Row: {
          category: string
          content: string
          created_at: string
          description: string | null
          example_inputs: Json
          id: string
          is_free: boolean
          title: string
          updated_at: string
          variables: Json
        }
        Insert: {
          category: string
          content: string
          created_at?: string
          description?: string | null
          example_inputs?: Json
          id?: string
          is_free?: boolean
          title: string
          updated_at?: string
          variables?: Json
        }
        Update: {
          category?: string
          content?: string
          created_at?: string
          description?: string | null
          example_inputs?: Json
          id?: string
          is_free?: boolean
          title?: string
          updated_at?: string
          variables?: Json
        }
        Relationships: []
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}